putting ripple color where it belongs
