+++ daemon/openrazer_daemon/misc/ripple_effect.py
@@ -4,11 +4,40 @@ Contains the functions and classes to perform ripple effects
 import datetime
 import logging
 import math
+import random
 import threading
 import time
 
 # pylint: disable=import-error
 from openrazer_daemon.keyboard import KeyboardColour
+
+COLOUR_CHOICES = (
+    (255, 0, 0),    # Red
+    (0, 255, 0),    # Green
+    (0, 0, 255),    # Blue
+    (255, 255, 0),  # Yellow
+    (0, 255, 255),  # Cyan
+    (255, 0, 255),  # Magenta
+)
+
+
+def random_colour_picker(last_choice, iterable):
+    """
+    Chose a random choice but not the last one
+
+    :param last_choice: Last choice
+    :type last_choice: object
+
+    :param iterable: Iterable object
+    :type iterable: iterable
+
+    :return: Choice
+    :rtype: object
+    """
+    result = random.choice(iterable)
+    while result == last_choice:
+        result = random.choice(iterable)
+    return result
 
 
 class RippleEffectThread(threading.Thread):
@@ -26,6 +55,7 @@ class RippleEffectThread(threading.Thread):
 
         self._colour = (0, 255, 0)
         self._refresh_rate = 0.040
+        self._last_colour_choice = None
 
         self._shutdown = False
         self._active = False
@@ -121,8 +151,11 @@ class RippleEffectThread(threading.Thread):
 
                 radiuses = []
 
-                for expire_time, (key_row, key_col), colour in self.key_list:
+                for expire_time, (key_row, key_col) in self.key_list:
                     event_time = expire_time - expire_diff
+
+                    colour = random_colour_picker(self._last_colour_choice, COLOUR_CHOICES)
+                    self._last_colour_choice = colour
 
                     now_diff = now - event_time
                     # Current radius is based off a time metric
